requirements.txt
pymongo[srv]
openai
python-dotenv